split schemas from db models
