+++ app/schemas/schema.py
@@ -18,7 +18,7 @@ from typing import Optional
 from pydantic import BaseModel
 from pydantic import Field
 
-from .base_models import APIResponse
+from .base import APIResponse
 
 
 class POSTSchema(BaseModel):
